[Fix] #70 - RateViewModel 삭제
- RateViewModel 삭제하고 RateViewModel에 있는 코드들 artistViewmodel로 이동
